Define tiled clustering example as a script used to benchmark CPU and GPU ops
- Script to execute and profile cebmf model fitting
- Define a new dependency group for profiling

+++ src/cebmf_torch/utils/mixture.py
@@ -141,7 +141,7 @@ def _calculate_scales(
     return torch.cat(
         [
             torch.tensor([0.0], device=device, dtype=dtype),
-            (1.0 / mult) ** (-seq.to(dtype=torch.float64)).to(dtype)
+            (1.0 / mult) ** (-seq.to(dtype=torch.float32)).to(dtype)
             * torch.tensor(sigmaamax, device=device, dtype=dtype),
         ]
     )
